make agnostic patterns go before enum + `FilterSkip`
"Agnostic" patterns (those not defined for any specific token) previously had to be added to some Unit token.
Now they are required to be added to the `Token` enum itself.

This also adds `FilterSkip` which either skips or returns an error, but never a token. It's useful for block comments.

Lastly, added some more documentation to the handbook.